pyproject.toml
[tool.poetry]
name = "talentscouter"
version = "0.1.0"
description = "A tool to analyze and rank CVs for job positions using AI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pypdf = "^3.17.1"
openai = "^1.12.0"
pandas = "^2.2.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
talentscouter = "src.main:main" 